requirements.txt
scipy==1.10.1
pyinterval==1.2.0
jupyterlab==3.6.3
matplotlib==3.7.1
mypy==1.2.0
